Move mapStepped to more appropriate place
